# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "CaseStatus": {
        "description": "Case lifecycle (workflow.md \u00a76.2). Terminal states have no outgoing edge.",
        "enum": [
          "DETECTED",
          "TRIAGED",
          "DIAGNOSING",
          "STRATEGY_FORMED",
          "AWAITING_APPROVAL",
          "EXECUTING",
          "MONITORING",
          "RECOVERED",
          "RESOLVED_ORGANIC",
          "EXPIRED",
          "SUPPRESSED",
          "REJECTED",
          "FAILED_PERMANENT"
        ],
        "title": "CaseStatus",
        "type": "string"
      },
...
      "ExperimentArm": {
        "description": "CONTROL cases receive no intervention at all, so that recovery can be\nreported as incremental lift rather than asserted (workflow.md \u00a714.2).",
        "enum": [
          "TREATMENT",
          "CONTROL"
        ],
        "title": "ExperimentArm",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Playbook": {
        "description": "The four revenue leaks (workflow.md \u00a75).",
        "enum": [
          "PAYMENT_FAILURE",
          "CHECKOUT_ABANDON",
          "RECEIVABLE",
          "SUBSCRIPTION"
        ],
        "title": "Playbook",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Autonomous revenue recovery agent for Razorpay merchants. Bounded agent: the LLM diagnoses and writes; it never touches money.",
    "title": "RevPilot AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cases": {
      "get": {
        "description": "Paginated case list.\n\n``arm`` is filterable specifically so a judge can ask for\n``?arm=CONTROL`` and see cases we deliberately did not act on. A control\narm nobody can inspect is indistinguishable from one that does not exist.",
        "operationId": "list_cases_api_v1_cases_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CaseStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "playbook",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Playbook"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Playbook"
            }
          },
          {
            "in": "query",
            "name": "arm",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExperimentArm"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Arm"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Cases Api V1 Cases Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List cases, filterable by status, playbook and arm",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}": {
      "get": {
        "description": "Everything that happened to one case, in order.\n\nAudit blocks are returned with their hashes so the trace can be checked\nagainst ``/audit/verify`` rather than taken on faith \u2014 the trace and the\nledger are the same events, and a discrepancy between them is exactly what\nthe chain exists to surface.",
        "operationId": "get_case_api_v1_cases__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Case Api V1 Cases  Case Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Full glass-box trace for one case",
        "tags": [
          "cases"
        ]
      }
    },
...
  }
}